requirements.txt
numpy==1.18.2
pandas==1.0.3
hdx-python-api==4.3.5